requirements.txt
flask
gunicorn
pytest
python-dotenv
flask_sqlalchemy
psycopg2-binary
flask_login


